requirements.txt
python-dotenv
streamlit
langchain 
langchain-community
langchain-core
langchain-google-genai
chromadb
pdfplumber